requirements.txt
fastapi
uvicorn[standard]
pymongo
python-frontmatter
watchdog
python-multipart
Jinja2
PyJWT
python-dotenv
python-multipart
passlib
bcrypt<4.1.0
autopep8